@theotherjimmy 2016-07-15
Would love to +1000, like @PrzemekWirkus but I would like to see these three functions de-duplicated. so only +999 ;P