@0xc0170 2016-07-15
@theotherjimmy yes, we could unify those in the Exporter class, as we can see they are the same for progen (same API calls). That shall come as separate PR